let user choose whether split windows 10 wim file

+++ tools/extract_win10_bootable_iso.sh
@@ -4,6 +4,7 @@ set -e
 
 path_to_iso=$1
 dest=$2
+is_split_wim=$3
 
 # let user choose path to iso
 if [ -z $path_to_iso ]
@@ -23,8 +24,14 @@ sudo mount $path_to_iso win_img -o loop
 sudo cp -r win_img/* tmp/
 
 # split install.wim to fit fat32 filesystem
-sudo wimsplit tmp/sources/install.wim tmp/sources/install.swm 2500
-sudo rm tmp/sources/install.wim
+if [ ! -z $is_split_wim ]
+then
+    if [ $is_split_wim = "y" ]
+    then
+	sudo wimsplit tmp/sources/install.wim tmp/sources/install.swm 2500
+	sudo rm tmp/sources/install.wim
+    fi
+fi
 
 # let user choose version of Windows 10 on install
 printf "[Channel]\r\nRetail\r\n" | sudo tee tmp/sources/ei.cfg
